Show current file and line number on Trac
Useful for sending someone a quick link to a WordPress function. Find
the point you want to show, and hit ^⌥⌘T… Safari will open to that file
and line number on Trac.

Caveat: Targets Trunk on Trac, not the current tag as I couldn't work
out a var which set the current WordPress version.