Show on Trac Browser command
Useful for sending someone a quick link to a WordPress function. Find
the point you want to show, and hit ^⌥⌘T… Safari will open to that file
and line number on Trac.

Auto-detects the version of WordPress in the project directory and the
WordPress directory using magic, mdfind, grep and regexes.